==> 01-scrape.ipynb <==
import re
import requests

pages = 106
pageuri = "https://mdc.mo.gov/conservation-areas-search?page=%d"

# We are searching for matches similar to <a href="/discover-nature/places/watson-access" hreflang="en">Watson Access</a>
matcher = re.compile(r'<a href="(/discover-nature/places/)([^"]+)" hreflang="en">([^<]+)</a>')
all_matches = []
for page in range(1, pages + 1):
    print(f"Page {page}")
    thispage = pageuri % (page)
    page = requests.get(thispage)
    for match in matcher.findall(page.text):
        print(f"  {match[2]}")
        all_matches.append(match[1:])

import pandas as pd
pd.DataFrame(all_matches, columns=['area_id', 'area_name']).to_csv('conservation_areas.tsv', sep='\t')

import pandas as pd
from biagen.scrape import fetch_url

areas_df = pd.read_csv('assets/conservation_areas.tsv', sep='\t')
place_url = 'https://mdc.mo.gov/discover-nature/places/%s'
lake_info = []
for i, row in areas_df.iterrows():
    print(i, row['area_name'])
    area_id = row['area_id']
    results = fetch_url(place_url % area_id, bare=True)
    # These are the last words in the header, so the end marks the beginning of the document
    start_index = results.find('Discover Nature Find Places To Go') + len('Discover Nature Find Places To Go')
    results = results[start_index:]
    lake_info.append(results)

# add the lake info to the lakes_df
areas_df['area_info'] = lake_info
areas_df.to_csv('assets/mo_conservation.tsv', sep='\t')
