==> 02-preprocess.ipynb <==
# Simple Entity Recognition

We are going to extract a few classes of information and filter our results down to 100 records.

from dotenv import load_dotenv
import logging
import pandas as pd

# Set up logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

load_dotenv()

# load some data
import html
import re
import pandas as pd
import numpy as np

from typing import Optional

mo_conservation = pd.read_csv('assets/mo_conservation.tsv.gz', sep='\t', index_col='id')

# Define a few regex pattern to identify size-related information
size_patterns = [
    re.compile(r'acres\s+(\d{1,3}(?:,\d{3})*(?:\.\d+)?)', re.IGNORECASE),
    re.compile(r'(\d{1,3}(?:,\d{3})*(?:\.\d+)?)[ -]acre', re.IGNORECASE),
]

def match_acres_area(info: str) -> Optional[float]:
    for s in size_patterns:
        match = s.search(info)
        if match:
            return float(match.group(1).replace(',', ''))
    return None

def match_location_type(title: str) -> str:
    if "Management Lands" in title:
        return "Management Lands"
    elif 'Towersite' in title:
        return 'Towersite'
    elif "Fish Hatchery" in title:
        return "Fish Hatchery"
    elif "Forestry Office" in title:
        return "Forestry Office"
    elif "Nature Center" in title:
        return "Nature Center"
    elif "Natural Area" in title:
        return "Natural Area"
    elif "Wildlife Area" in title:
        return "Wildlife Area"
    elif "Park" in title:
        return "Park"
    elif "Pond" in title:
        return "Pond"
    elif "Headquarters" in title:
        return "Headquarters"
    elif "Regional Office" in title:
        return "Regional Office"
    elif "Range" in title:
        return "Range"
    elif "Office" in title:
        return "Office"
    elif "Discovery Center" in title:
        return "Discovery Center"
    elif "Education Area" in title:
        return "Education Area"
    elif "Education Center" in title:
        return "Education Center"
    elif "Service Center" in title:
        return "Service Center"
    elif "Prairie" in title:
        return "Prairie"
    elif "Reservation" in title:
        return "Reservation"
    elif "Blue Hole" in title:
        return "Blue Hole"
    elif "Reservoir" in title:
        return "Reservoir"
    elif "Forest" in title:
        return "Forest"
    elif "Bend" in title:
        return "Bend"
    elif "Nursery" in title:
        return "Nursery"
    elif "Preserve" in title:
        return "Preserve"
    elif "Tract" in title:
        return "Tract"
    elif 'Access' in title:
        return 'Access'
    elif 'Conservation Area' in title:
        return 'Conservation Area'
    elif 'Lake' in title:
        return 'Lake'
    elif "Memorial" in title:
        return "Memorial"
    else:
        return 'Unknown'

def fix_html(value: str) -> str:
    """
    Decodes HTML entities in a name string, converting them to their proper characters.

    Args:
        value (str): The name string potentially containing HTML entities.

    Returns:
        str: The decoded name string.
    """
    # Decode HTML entities such as &#039; into proper characters
    return html.unescape(value)

# Fix our names
mo_conservation['area_name'] = mo_conservation['area_name'].apply(fix_html)

# Apply the function to extract size information
mo_conservation['acres'] = mo_conservation['area_info'].apply(match_acres_area)
mo_conservation['location_type'] = mo_conservation['area_name'].apply(match_location_type)

print(f"{mo_conservation['acres'].isna().sum()} missing acres")
for r in mo_conservation[mo_conservation['acres'].isna() ]['area_info']:
    print(r)
    
print(f"{(mo_conservation['location_type'] == "Unknown").sum()} unknown locations")
for r in mo_conservation[mo_conservation['location_type'] == "Unknown"]['area_name']:
    print(r)

mo_conservation.value_counts('location_type')

def sample_groups(location_type: str) -> int:
    return {
        'Access': 0,
        'Bend': 0,
        'Blue Hole': 2,
        'Conservation Area': 0,
        'Discovery Center': 3,
        'Education Area': 3,
        'Education Center': 3,
        'Fish Hatchery': 3,
        'Forest': 0,
        'Forestry Office': 1,
        'Headquarters': 1,
        'Lake': 0,
        'Management Lands': 2,
        'Memorial': 3,
        'Natural Area': 0,
        'Nature Center': 3,
        'Office': 1,
        'Park': 0,
        'Pond': 0,
        'Prairie': 0,
        'Preserve': 4,
        'Range': 3,
        'Regional Office': 1,
        'Reservation': 0,
        'Reservoir': 0,
        'Service Center': 1,
        'Towersite': 5,
        'Tract': 0,
        'Wildlife Area': 4,
    }.get(location_type, -1)

mo_conservation['sample_group'] = mo_conservation['location_type'].apply(sample_groups)

group_names = {
    1: 'Administration',
    2: 'Natural Areas',
    3: 'Education',
    4: 'Wildlife',
    5: 'Towersites',
}

mo_conservation['category'] = "Missouri Conservation Areas"
mo_conservation['subcategory'] = mo_conservation['sample_group'].apply(lambda x: group_names.get(x, 'Unclassified'))

sample_mask = mo_conservation['sample_group'] > 0
print(f"{sample_mask.sum()} in sample pool")
mo_conservation.value_counts('sample_group')

sample_pool = mo_conservation[mo_conservation['sample_group'] > 0].reset_index(drop=True).copy()
sample_pool['selected'] = 0
sample_pool.head()

import numpy as np

pd.set_option('display.max_rows', 5000)

# for each of our 5 bins, we will get the indices of all of the elements, and then randomly select 20, and use the resulting set to update our index
# setting the value of selected to 1

# clear the 'selected' column if not already present
sample_pool['selected'] = 0

# For each of our 5 bins, we will get the indices of all elements, randomly select 20, 
# and use the resulting set to update our DataFrame by setting the value of 'selected' to 1
need = 20
for i in range(1, 6):
    candidates = sample_pool[sample_pool['sample_group'] == i].index.values

    n = min(need, len(candidates))
    if n < 20:
        print(f"Not enough candidates ({n}) for group {i}, overflowing to next group")
        need = 20 + (20 - n)
    else:
        need = 20

    # Using numpy's random choice, we get our selections
    selected_indices = np.random.choice(candidates, n, replace=False)

    # Now with a pandas location indexer, we set our selected values
    sample_pool.loc[selected_indices, 'selected'] = 1

result_df = sample_pool[sample_pool['selected'] == 1]\
    .sort_values(['category', 'subcategory', 'location_type', 'area_name'])\
    .reset_index(drop=True)\
    [['area_id', 'area_name', 'category', 'subcategory', 'location_type', 'area_info']]\
    .copy()
result_df.to_csv('assets/random_100_areas.tsv', sep='\t', index_label='id')
result_df


pd.set_option('display.max_rows', 5000)
lakes = mo_conservation[
    mo_conservation['location_type'] == 'Lake'
    ]\
    .sort_values('acres', ascending=False)\
    .reset_index(drop=True)[['area_id', 'area_name', 'location_type', 'acres', 'area_info']]
lakes[:100].to_csv('assets/top_100_lakes.tsv', sep='\t', index_label='id')
lakes = pd.read_csv('assets/top_100_lakes.tsv', sep='\t', index_col='id')
lakes
