==> 03-process_one.ipynb <==
# Abstractive Summarization

This notebook contains a sample for abstractive summarization using chain of density prompting.

https://arxiv.org/abs/2309.04269

Griffin Adams et al. introduce "Chain of Density" (CoD), an iterative GPT-4 prompting method generating increasingly entity-dense summaries at fixed length. CoD identifies 1-3 missing salient entities, incorporating them over five steps. Tested on 100 CNN/DailyMail articles, human evaluation (by the first four authors) reveals preference for summaries with entity density (entities/token) ~0.15, surpassing vanilla GPT-4 (0.122) and matching human-written (0.151) summaries. Entity density increases from 0.089 (step 1) to 0.167 (step 5), with step 3 (0.148) most preferred (expected step: 3.06). Low inter-annotator agreement (Fleiss' kappa: 0.112) indicates task subjectivity. Abstractiveness (extractive density), fusion (relative ROUGE gain), and lead bias measured; NLTK/Spacy used for tokenization/entity recognition. GPT-4 Likert-scale assessments show informativeness peaks at step 4, while Quality/Coherence decline after steps 2/1. Qualitative analysis reveals coherence/informativeness trade-off, with factual correctness challenging at high densities. CoD outperforms previous entity-based summarization approaches. Study limitations: news-only focus, evaluation subjectivity. 500 annotated/5,000 unannotated CoD summaries open-sourced for further research, enabling density distillation into open-source models.

This particular summary was generation 3 of Claude Sonnet 3.5 summarizing the text of the PDF using `Summarize this PDF.` and `Identify 10 key missing points, and any errors in the summary, and output a new, improved summary. Increase the density of the resulting summary, but do not increase the length.`

from dotenv import load_dotenv
import logging
import pandas as pd

# Set up logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

load_dotenv()

Step 1: Pick your model provider

from biagen.llm import CohereProvider, GroqProvider, AIStudioProvider

llm = AIStudioProvider.from_env()

Step 2: Load some data

# load some data

area_df = pd.read_csv('assets/mo_conservation.tsv.gz', sep='\t')

data = area_df.loc[432]

article = data['area_info']

data

summaries = []

prompt_a = f"""Analyze the purpose and main topics of this article. Provide a detailed and accurate analysis spanning all covered topics.

{article}

## Analysis (the most significant topics, with one sentence commentary)
"""

analysis = llm.generate_one(prompt_a, max_tokens=1024, temperature=0.7, stop_sequences=None)

print(analysis)
len(analysis.split())

prompt_s = f"""Given the article and the following analysis, provide a detailed, erudite, succinct, and accurate summary.

## Article

```article
{article}
```

## Analysis

{analysis}

## Summary (5 paragraphs)

"""

summary= llm.generate_one(prompt_s, max_tokens=1024, temperature=0.5, stop_sequences=None)

summaries.append(summary)

print(summary)
len(summary.split())

### Improvement Loop

# Get our last summary

summary = summaries[-1]

# Check for missing information in the summary

prompt_m = f"""Given the article and the current summary, identify missing information or ways to improve the summary.

## Article

```article
{article}
```

## Current Summary

{summary}

## Missing Information (provide 12 entries of novel information not contained in the summary)
"""

missing = llm.generate_one(prompt_m, max_tokens=1024, temperature=0.7, stop_sequences=None)

print(missing)
len(missing.split())

prompt_r = f"""Using the current summary and the identified missing information, identify what information is important and unimportant.

## Article

```article
{article}
```

## Current Summary

{summary}

## Missing Information

{missing}

## Most Important Information
"""

important = llm.generate_one(prompt_r, max_tokens=2048, temperature=0.5, stop_sequences=None)

print(important)
len(important.split())

prompt_r = f"""Using the current summary and the identified missing information, create an improved summary of the article.

## Article

```article
{article}
```

## Current Summary

{summary}

## Missing Information

{missing}

## Ideas To Consider

{important}

## Improved Summary (reorganized, expanded, and including new information; 6 paragraphs)
"""

resummarized = llm.generate_one(prompt_r, max_tokens=2048, temperature=0.5, stop_sequences=None)

summaries.append(resummarized)

print(resummarized)
len(resummarized.split())
